app with user Authentication

+++ backend/controller/expense.js
@@ -3,14 +3,15 @@ const Expense = require('../model/expense');
 
 exports.newExpense = async (req, res, next) => {
   try {
-    description = req.body.description;
-    price = req.body.price;
-    category = req.body.category;
+    let description = req.body.description;
+    let price = req.body.price;
+    let category = req.body.category;
 
     let newExpense = await Expense.create({
       Description: description,
       Price: price,
-      Category: category
+      Category: category,
+      userId: req.user.id
     })
     res.status(200).send(newExpense);
 
@@ -20,18 +21,45 @@ exports.newExpense = async (req, res, next) => {
 }
 
 exports.allExpense = (req, res, next) => {
-  Expense.findAll()
-    .then(result => res.send(result))
+  let Id = req.user.id;
+  // console.log(req.user.id)
+  Expense.findAll({ where: { userId: Id } })
+    .then(result => {
+      // console.log(result)
+      res.send(result)
+    })
     .catch(err => console.log(err))
 }
 
 exports.deleteExpense = async (req, res, next) => {
   try {
     let Id = req.body.id;
+    let userId = req.user.id;
     // console.log(Id);
+    // console.log(userId);
 
-    await Expense.destroy({ where: { id: Id } });
+    await Expense.destroy({ where: { id: Id, userId: userId } });
     res.send('item deleted');
+  }
+  catch (err) {
+    console.log(err);
+  }
+}
+
+
+exports.updateExpense = async (req, res, next) => {
+  try {
+    let Id = req.body.id;
+    let userId = req.user.id;
+
+    await Expense.update(
+      {
+        Description: req.body.description,
+        Price: req.body.price,
+        Category: req.body.category,
+      },
+      { where: { id: Id, userId: userId } });
+    res.send('item updeted');
   }
   catch (err) {
     console.log(err);
